Removed JobFile option because it doesn't make sense.
DART does not currently export individual jobs, and even if it did, the export
would require substantial reformatting of the Job JSON to be usable by
dart-runner. It doesn't make sense anyway to export single jobs to be run on
a server.

Since dart-runner now handles job params in conjunction with workflows, this
will be the means of executing one-off jobs in dart-runner.